# 8302191: Performance degradation for float/double modulo on Linux