Backport of [JDK-8302191](https://bugs.openjdk.java.net/browse/JDK-8302191). Applies cleanly. It is already committed in trunk.

```
build/linux-x86_64-server-fastdebug/jdk/bin/java -jar ~/azul/jmh-jdk-microbenchmarks-git/micros-jdk8/target/micros-jdk8-1.0-SNAPSHOT.jar org.openjdk.bench.java.lang.FloatDoubleRem -i 2 -r 2 -wi 2 -f 2
unpatched:
Benchmark                            Mode  Cnt    Score    Error  Units
FloatDoubleRem.calcDoubleJava        avgt    4  106.225 ± 29.934  ns/op
FloatDoubleRem.calcFloatJava         avgt    4  101.977 ±  8.528  ns/op
FloatDoubleRem.cornercaseDoubleJava  avgt    4   10.052 ±  0.132  ns/op
FloatDoubleRem.cornercaseFloatJava   avgt    4    7.336 ±  0.798  ns/op
patched:
Benchmark                            Mode  Cnt   Score   Error  Units
FloatDoubleRem.calcDoubleJava        avgt    4  31.701 ± 7.517  ns/op
FloatDoubleRem.calcFloatJava         avgt    4  33.873 ± 1.650  ns/op
FloatDoubleRem.cornercaseDoubleJava  avgt    4  10.108 ± 0.657  ns/op
FloatDoubleRem.cornercaseFloatJava   avgt    4  10.141 ± 0.466  ns/op
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8302191](https://bugs.openjdk.org/browse/JDK-8302191): Performance degradation for float/double modulo on Linux


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk20u.git pull/46/head:pull/46` \
`$ git checkout pull/46`

Update a local copy of the PR: \
`$ git checkout pull/46` \
`$ git pull https://git.openjdk.org/jdk20u.git pull/46/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 46`

View PR using the GUI difftool: \
`$ git pr show -t 46`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk20u/pull/46.diff">https://git.openjdk.org/jdk20u/pull/46.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk20u/pull/46#issuecomment-1494449691)